requirements.txt
flask==2.3.2
face_recognition==1.3.0
dlib==19.24.6
Pillow==10.4.0
opencv-python==4.10.0.84
mysql-connector-python==8.0.33
numpy==1.23.5

